+++ contracts/flow/basic/Flow.sol
@@ -16,20 +16,21 @@ contract Flow is ReentrancyGuard, FlowCommon {
     event Initialize(address sender, FlowConfig config);
 
     /// @param config_ allowed flows set at initialization.
-    function initialize(FlowConfig calldata config_)
-        external
-        initializer
-    {
+    function initialize(FlowConfig calldata config_) external initializer {
         __FlowCommon_init(config_.flowConfig);
         emit Initialize(msg.sender, config_);
     }
 
-    function _previewFlow(address flow_, uint id_, SignedContext[] memory signedContexts_)
-        internal
-        view
-        returns (FlowTransfer memory)
-    {
-        (StackTop stackBottom_, StackTop stackTop_) = flowStack(flow_, id_, signedContexts_);
+    function _previewFlow(
+        address flow_,
+        uint256 id_,
+        SignedContext[] memory signedContexts_
+    ) internal view returns (FlowTransfer memory) {
+        (StackTop stackBottom_, StackTop stackTop_) = flowStack(
+            flow_,
+            id_,
+            signedContexts_
+        );
         return LibFlow.stackToFlow(stackBottom_, stackTop_);
     }
 
@@ -46,7 +47,11 @@ contract Flow is ReentrancyGuard, FlowCommon {
         uint256 id_,
         SignedContext[] memory signedContexts_
     ) external payable virtual nonReentrant returns (FlowTransfer memory) {
-        FlowTransfer memory flowTransfer_ = _previewFlow(flow_, id_, signedContexts_);
+        FlowTransfer memory flowTransfer_ = _previewFlow(
+            flow_,
+            id_,
+            signedContexts_
+        );
         registerFlowTime(_flowContextScratches[flow_], flow_, id_);
         return LibFlow.flow(flowTransfer_, address(this), payable(msg.sender));
     }
