+++ contracts/interpreter/run/RainInterpreter.sol
@@ -89,8 +89,11 @@ abstract contract RainInterpreter {
         internal
         view
         returns (StackTop);
-    using LibCast for function(InterpreterState memory, Operand, StackTop) internal view returns (StackTop)[];
-    using LibConvert for uint[];
+    using LibCast for function(InterpreterState memory, Operand, StackTop)
+        internal
+        view
+        returns (StackTop)[];
+    using LibConvert for uint256[];
     using LibInterpreterState for StateConfig;
 
     /// Default is to disallow all storage access to opcodes.
@@ -146,7 +149,9 @@ abstract contract RainInterpreter {
                     scratch_,
                     contextScratch_,
                     stackLength_,
-                    opcodeFunctionPointers().asUint256Array().unsafeTo16BitBytes()
+                    opcodeFunctionPointers()
+                        .asUint256Array()
+                        .unsafeTo16BitBytes()
                 );
         }
     }
