+++ contracts/interpreter/shared/RainterpreterV1.sol
@@ -8,15 +8,19 @@ contract RainterpreterV1 is IInterpreter, RainInterpreter {
     using LibStackTop for StackTop;
     using LibInterpreterState for bytes;
     using LibInterpreterState for InterpreterState;
-    using LibCast for function(InterpreterState memory, Operand, StackTop) view returns (StackTop)[];
-    using LibConvert for uint[];
+    using LibCast for function(InterpreterState memory, Operand, StackTop)
+        view
+        returns (StackTop)[];
+    using LibConvert for uint256[];
 
     function eval(
         address statePointer_,
         SourceIndex entrypoint_,
         uint256[][] memory context_
     ) external view returns (uint256[] memory) {
-        InterpreterState memory state_ = SSTORE2.read(statePointer_).deserialize();
+        InterpreterState memory state_ = SSTORE2
+            .read(statePointer_)
+            .deserialize();
         state_.context = context_;
         StackTop stackTop_ = state_.eval(entrypoint_);
         uint256 stackLength_ = state_.stackBottom.toIndex(stackTop_);
@@ -25,7 +29,9 @@ contract RainterpreterV1 is IInterpreter, RainInterpreter {
     }
 
     function functionPointers() external view returns (bytes memory) {
-        return opcodeFunctionPointers().asUint256Array().unsafeTo16BitBytes();
+        return opcodeFunctionPointers()
+            .asUint256Array()
+            .unsafeTo16BitBytes();
     }
 
     function localEvalFunctionPointers()
